fix(webserver): update display options
The new `roles` field was included in the context. Since every QuestionDisplayOptions field gets a true/false selection but `roles` expects a set, trying to update display options resulted in an error.

This commit excludes the roles from the template entirely. In future we should allow choosing the user role and make the template less fragile.